requirement.txt
agate==1.9.1
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.2.0
attrs==25.4.0
babel==2.18.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
daff==1.4.2
dbt-adapters==1.22.10
dbt-common==1.37.3
dbt-core==1.11.7
dbt-extractor==0.6.0
dbt-postgres==1.10.0
dbt-protos==1.0.443
dbt-semantic-interfaces==0.9.0
debugpy==1.8.20
decorator==5.2.1
deepdiff==8.6.2
defusedxml==0.7.1
executing==2.2.1
fastjsonschema==2.21.2
fqdn==1.5.1
googlemaps==4.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.9.0
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isodate==0.7.2
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
lark==1.3.1
leather==0.4.1
MarkupSafe==3.0.3
mashumaro==3.14
matplotlib-inline==0.2.1
mistune==3.2.0
more-itertools==10.8.0
msgpack==1.1.2
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook==7.5.5
notebook_shim==0.2.4
numpy==2.3.4
orderly-set==5.5.0
packaging==26.0
pandas==2.3.3
pandocfilters==1.5.1
parsedatetime==2.6
parso==0.8.6
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.9.4
prometheus_client==0.24.1
prompt_toolkit==3.0.52
protobuf==6.33.6
psutil==7.2.2
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.0.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.3
redis==7.4.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
Send2Trash==2.1.0
setuptools==82.0.1
six==1.17.0
snowplow-tracker==1.1.0
soupsieve==2.8.3
sqlparse==0.5.4
stack-data==0.6.3
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.4.0
tornado==6.5.5
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
widgetsnbextension==4.0.15
youtube-transcript-api==1.2.3
yt-dlp==2025.10.14
zipp==3.23.0
